유저 홈 화면 - UserDesignerProfileActivity.kt, AllPricesFragment.kt : 가격 일부, 가격 전체보기 구현

+++ app/src/main/java/com/blooburn/owere/util/DesignerProfileHandler.kt
@@ -11,7 +11,7 @@ interface DesignerProfileHandler {
     /**
      * 평점 -> 별점 변환
      */
-    fun convertRatingToStar(rating: Float): String {
+    fun convertRatingToStar(rating: Double): String {
         return when{
             rating < 0 -> ""
             0 <= rating && rating < 1  -> "☆"
